add caching for posts data using Redis

+++ post-service/src/controllers/post.controller.js
@@ -2,6 +2,14 @@ const logger = require('../utils/logger')
 const Post = require('../models/post')
 const { validateCreatePost } = require('../utils/validation')
 
+async function invalidatePostsCache(req, postId) {
+  const cachedKey = `post:${postId?.toString()}`
+  await req.redisClient.del(cachedKey)
+
+  const keys = await req.redisClient.keys('posts:*')
+  if (keys.length > 0) await req.redisClient.del(keys)
+}
+
 const createPost = async (req, res) => {
   logger.info('Create post endpoint hit...')
 
@@ -19,6 +27,7 @@ const createPost = async (req, res) => {
 
     const newlyCreatedPost = new Post({ user: userId, mediaIds, content })
     await newlyCreatedPost.save()
+    await invalidatePostsCache(req, newlyCreatedPost._id)
 
     logger.info('Post created successfully.', newlyCreatedPost)
     res.status(201).json({ success: true, message: 'Post created successfully.' })
@@ -33,8 +42,16 @@ const getById = async (req, res) => {
 
   try {
     const { id } = req.params
+    const cachedkey = `post:${id}`
+    const cachedPost = await req.redisClient.get(cachedkey)
+
+    if (cachedPost) {
+      logger.info('Loaded data from Redis cache.')
+      return res.json(JSON.parse(cachedPost))
+    }
 
     const post = await Post.findById(id).lean()
+    await req.redisClient.setex(cachedkey, 300, JSON.stringify(post))
     res.json(post)
   } catch (error) {
     logger.error('Error fetching post', error)
@@ -50,6 +67,13 @@ const getPosts = async (req, res) => {
     const limit = parseInt(req.query.limit) || 10
     const offset = (page - 1) * limit
 
+    const cachedKey = `posts:${page}:${limit}`
+    const cachedPosts = await req.redisClient.get(cachedKey)
+    if (cachedPosts) {
+      logger.info('Loaded data from Redis cache.')
+      return res.json(JSON.parse(cachedPosts))
+    }
+
     const posts = await Post.find().sort({ _id: -1 }).skip(offset).limit(limit).lean()
     const total = await Post.countDocuments()
 
@@ -59,6 +83,9 @@ const getPosts = async (req, res) => {
       totalPages: Math.ceil(total / limit),
       total,
     }
+
+    // setex = SET with EXpire
+    await req.redisClient.setex(cachedKey, 300, JSON.stringify(result))
 
     res.json(result)
   } catch (error) {
@@ -72,7 +99,8 @@ const deleteById = async (req, res) => {
 
   try {
     const { id } = req.params
-    await Post.deleteOne({ _id: id})
+    await Post.deleteOne({ _id: id, user: req.user.userId })
+    await invalidatePostsCache(req, id)
     res.json({ success: true, message: 'Deleted successfully.' })
   } catch (error) {
     logger.error('Error deleting posts', error)
